web-server/requirements.txt
fastapi
uvicorn
requests
httpx
python-jose